Fix MoveReplicaSuggester Comparator
Previously, the old comparator that was used in the MoveReplicaSuggester
was causing "Comparison method violates its general contract!" our
theory was due to the fact it was possible for both replicas to be
leader which was causing the TimSort to spit out the exception.

This new comparator will accept both replicas being leader and will also
sort the replicas alphabetically between the two.

Finally there is a unit test just to validate that given a list of
Replica's, it'll sort it accordingly.

This has also been tested on my dev cluster numerous of times, and no
longer getting the exception

Made suggested changes