Fix termfreqfreq for non-distributed searches
We were relying on the FacetMerger to finalise the termfreqfreq results
(converting the raw frequencies into the frequency-of-frequencies), but
this is only invoked during distributed search - e.g. it will be skipped
when we are only searching a single collection/shard.

To address this, we need to modify the SlotAcc to check whether we are
in distributed mode before serializing each shard's results - when there
is only a single shard, we serialize the results into their final form,
rather than the intermediate form required for distributed search.