Cache IndexSchema on SchemaSimilarityFactory to avoid NPE with shared schemas
Using <bool name="shareSchema">true</bool> means that a single
SchemaSimilarityFactory would be shared between multiple cores.
Therefore inform(SolrCore) would be called multiple times.  This leads
to a race condition when the Similarity object is being used for query
at the same time as a new core is being loaded.  This results in a
NullPointerException as the new SolrCore is not yet fully initialised.