Fix UpsertCondition due to Solr 7 encoding of String types in javabin format.
This relates to a change in behaviour introduced in SOLR-12983 which affected how javabin encodes
strings, they are ByteArrayUtf8CharSequence and not String objects (but both are CharSequence). A
similar problem with UpdateRequestProcessors was noted and fixed in SOLR-13255 but that did not
fix the case of Atomic Updates where the field is a Map from operation->value and hence the value
was not a String anymore, but a ByteArrayUtf8CharSequence.